+++ src/utils.rs
@@ -166,6 +166,6 @@ impl FromStr for Fraction {
     }
 }
 
-pub fn poison_damage(severity: u8) -> u32 {
-    (severity as u32 + 4) / 5
+pub fn poison_damage(severity: u32) -> u32 {
+    (severity + 4) / 5
 }
